requirements.txt
# Main dependencies
pyyaml>=6.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
